Create sample service script
This is the main script that uses a combination of all the submodule
components.  It generates Franca IDL from VSS (using a specified part
of the tree which is defined in the script now), then generates
CommonAPI binding code from the Franca IDL, then creates a cmake project
for the SOME/IP variant of generation, and compiles it into an
executable (in the process linking to the various vSomeip and CommonAPI
runtime libraries).

It generates D-Bus code to test the function, but the sample executable
is based on SOME/IP

To use this fully, modify main.cpp and add any other required files
to create the actual program or SOME/IP based service. That code would
refer to (inherit classes etc) from the auto-generated proxies and
stubs.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>